cart controller updated

+++ E-Commerce/src/main/java/com/masai/beans/Item.java
@@ -42,7 +42,7 @@ public class Item {
 	private Product product;
 	
 	@JsonIgnore
-	private Integer itemPrice;
+	private Double itemPrice;
 	
 	
 	private Integer requiredQuantity;
